% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{country_names}
\alias{country_names}
\title{Country names}
\format{
\subsection{\code{country_names}}{

A character vector of country names (length = 231 countries)
}
}
\source{
https://gadm.org
}
\usage{
country_names
}
\description{
Country names
}
\keyword{datasets}
