% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.vitripennis_pilkington2014.R
\docType{data}
\name{h.vitripennis_pilkington2014}
\alias{h.vitripennis_pilkington2014}
\title{\if{html}{\out{<b>}}\if{html}{\out{<i>}}Homalodisca vitripenis\if{html}{\out{</i>}} whole life cycle development rate across temperatures\if{html}{\out{</b>}}}
\format{
\subsection{\code{h.vitripennis_pilkington2014}}{

A data frame with 357 rows and 4 columns. This data.frame is a simulated reconstruction performed on pooled, summarized data in the source article
(Pilkington et al. 2014) using simulations under normal distribution based on mean estimate,
standard errors and sample sizes, as suggested by \if{html}{\out{<a href="https://doi.org/10.1002/jrsm.1331">}}Papadimitropoulou et al. (2019)\if{html}{\out{</a>}}. \if{html}{\out{</br>}}

The workflow is reproducible and available in \verb{/data-raw} folder of the \if{html}{\out{<a href="https://github.com/EcologyR/mappestRisk">}}mappestRisk GitHub repository\if{html}{\out{</a>}},
which includes both the original summarized data set -\code{homalodisca_vitripennis_pilkington2014.xlsx}- and the
R script with the simulation procedures in \code{clean&sim_devdata_homalodisca_vitripennis.R}.
.

\describe{
\item{reference}{"Pilkington2014" refers to the source paper as cited below in section \code{Source}.}
\item{temperature}{Temperature treatments (ºC).}
\item{rate_development}{Rate of Development (1/days), i.e. days necessary to fulfill development of all life stages from egg to adult (see \code{stage})}
\item{stage}{Life stage or instar evaluated. In this case, only data of the whole life cycle (from egg to adult) were used}
}
}
}
\source{
{\if{html}{\out{<b>}}Pilkington, L., Lewis, M., Jeske, D. and Hoddle, M. (2014)\if{html}{\out{</b>}}. Calculation and Thematic Mapping of Demographic Parameters
for \if{html}{\out{<i>}}Homalodisca vitripennis\if{html}{\out{</i>}} (Hemiptera: Cicadellidae) in California.
Ann. Entomol. Soc. Am. 107: 424-434. \url{https://doi.org/10.1603/AN13144}}.\if{html}{\out{</br>}}

Licence: \if{html}{\out{<a href="https://creativecommons.org/licenses/by-nc/3.0">}}CC BY-NC 3.0\if{html}{\out{</a>}} (modified material).}
}
\usage{
data(h.vitripennis_pilkington2014)
}
\description{
A reconstructed data set from Pilkington et al. (2014) on days of development across different constant temperatures and life stages
Scientific Article
}
\keyword{datasets}
