% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_map.R
\name{interactive_map}
\alias{interactive_map}
\title{Generate an interactive map from a risk map}
\usage{
interactive_map(
  x,
  legend_title = NULL,
  map_type = c("high", "medium"),
  path_out = NULL,
  pal = NULL
)
}
\arguments{
\item{x}{a SpatRaster object}

\item{legend_title}{the title of the legend.}

\item{map_type}{the type of map. One of "high" or "medium" risk.}

\item{path_out}{(optional) the name and the path of the output file. Example: "data/map.html"}

\item{pal}{the palette used to plot the risk map. Default ...}
}
\value{
a dynamic map and, optionally, a html with the dynamic map
}
\description{
Generate an interactive map from a risk map
}
\examples{
\dontrun{
r <- system.file("extdata", "trioza_rast.tif", package = "mappestRisk")
# r <- terra::rast("inst/extdata/trioza_rast.tif")
interactive_map(x = r, map_type = "high", path_out = paste0(tempdir(), "test_map.html"))
}

}
