% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devmodels.R
\name{plot_devmodels}
\alias{plot_devmodels}
\title{Explore shape of fitted thermal performance curves to choose an appropriate model based on both statistics and ecological sense}
\usage{
plot_devmodels(temp, dev_rate, fitted_parameters)
}
\arguments{
\item{temp}{a vector containing temperature treatments (predictor variable), must have at least three different temperature treatments}

\item{dev_rate}{a vector containing development rate estimates (1/days of development); must be of same length than temp}

\item{fitted_parameters}{a tibbl obtained with fit_models() function including parameter names, estimates, se and AICs}
}
\value{
a ggplot with facets printing development rate data provided by the
user and the predicted values by the models fitted with fit_devmodels
}
\description{
Explore shape of fitted thermal performance curves to choose an appropriate model based on both statistics and ecological sense
}
\examples{
  a.citricidus_tsai1999 <- readRDS("data/a.citricidus_tsai1999.rds")
source("R/dario_model_fitting.R")
plot_devmodels(temp = p.xylostella_liu2002$temperature,
               dev_rate = p.xylostella_liu2002$rate_development,
               fitted_parameters = cabbage_moth_fitted)

}
