% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_curves.R
\name{predict_curves}
\alias{predict_curves}
\title{Propagate parameter uncertainty of TPC fits using bootstrap with residual resampling}
\usage{
predict_curves(
  temp,
  dev_rate,
  fitted_parameters,
  model_name_2boot,
  propagate_uncertainty = TRUE,
  n_boots_samples = 100
)
}
\arguments{
\item{temp}{a vector containing temperature treatments (predictor variable).
It must have at least four different temperature treatments. It must be numeric
and not containing NAs.}

\item{dev_rate}{a vector containing development rate estimates, calculated as
the reciprocal of days of development at each temperature (i.e., 1/days of development).
It must be numeric and of same length as \code{temp}.}

\item{fitted_parameters}{a \code{tibble} obtained with \code{fit_devmodels()} function including parameter names,
estimates, se, AICs and \if{html}{\out{<nls>}} objects (i.e. fitted_models) from \code{fit_devmodels()}
under \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}} approach.}

\item{model_name_2boot}{one or several TPC models of those fitted in \code{fit_devmodels()}.
\code{model_name_2boot = "all"} is not allowed in this function, as bootstrapping procedure is quite slow.
We recommend to apply this function only to a small pre-selection of models (e.g., one to four) based on statistical and ecological
criteria with help from \code{plot_devmodels()} function.}

\item{propagate_uncertainty}{Whether to propagate parameter uncertainty by bootstrap with residual resampling or not.
If \code{FALSE}, the function will return the predictions from the fitted TPCs for the chosen models. If \code{TRUE}, bootstrap is applied
using residual resampling to obtain multiple TPCs using \code{\link[car:Boot]{car::Boot()}} as detailed in vignettes of \code{rTPC} package.
Default to \code{TRUE}.}

\item{n_boots_samples}{number of bootstrap resampling iterations (def. 100). A larger number
of iterations makes the resampling procedure more robust, but 100 is usually
enough for propagating parameter uncertainty, as increasing \code{n_boots_samples} will increase
computation time for predicting resampled TPCs.}
}
\value{
a \code{tibble} object with as many curves -TPCs- as the number of iterations provided
in \code{n_boots_samples} argument if \code{propagate_uncertainty = TRUE}. Otherwise, it will return just one prediction TPC
from model fit estimates. Each resampled TPC will consist of a collection of predictions for
a set of temperatures from \code{temp - 20} to \code{temp + 15} with a resolution of 0.1ºC and a unique identifier
called \code{iter}. In addition to the uncertainty TPCs, the estimate TPC is also explicit in the output tibble.
}
\description{
Propagate parameter uncertainty of TPC fits using bootstrap with residual resampling
}
\examples{
data("b.schwartzi_satar2002")

fitted_tpcs_bschwartzi <- fit_devmodels(temp = b.schwartzi_satar2002$temperature,
                                      dev_rate = b.schwartzi_satar2002$rate_value,
                                      model_name = "all")

plot_devmodels(temp = b.schwartzi_satar2002$temperature,
               dev_rate = b.schwartzi_satar2002$rate_value,
               fitted_parameters = fitted_tpcs_bschwartzi,
               species = "Brachycaudus swartzi",
               life_stage = "Nymphs") #choose "briere2", "thomas" and "lactin2"

#3. Obtain prediction TPCs with bootstraps for propagating uncertainty:
tpc_preds_boots_bschwartzi <- predict_curves(temp = b.swartzi_satar2002$temperature,
                                                      dev_rate = b.swartzi_satar2002$rate_value,
                                                      fitted_parameters = fitted_tpcs_bswartzi,
                                                      model_name_2boot = c("briere2", "thomas", "lactin2"),
                                                      propagate_uncertainty = TRUE,
                                                      n_boots_samples = 100)


head(tpc_preds_boots_bschwartzi)

}
\seealso{
\code{browseVignettes("rTPC")} for model names, start values searching workflows and
bootstrapping procedures using both \code{\link[rTPC:get_start_vals]{rTPC::get_start_vals()}} and \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}
}
