% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_suitability_bounds.R
\name{thermal_suitability_bounds}
\alias{thermal_suitability_bounds}
\title{Obtain thermal traits representing boundaries of optimal thermal suitability for pest risk occurrence assessment}
\usage{
thermal_suitability_bounds(
  fitted_parameters,
  model_name,
  suitability_threshold = NULL
)
}
\arguments{
\item{fitted_parameters}{a tibble/data.frame obtained with fit_models() function}

\item{model_name}{one or several of the models in fitted_parameters.
Must belong to available models in list_available_models}

\item{suitability_threshold}{Quantile (\%) of the curve; e.g. 75 would give
the temperatures within the values at which development rate is in the
top 25\% of its possible values}
}
\value{
a tibble/data.frame with name of the model and two thermal traits representing the above-mentioned suitability boundaries
}
\description{
Obtain thermal traits representing boundaries of optimal thermal suitability for pest risk occurrence assessment
}
\examples{
 a.citricidus_tsai1999 <- readRDS("data/a.citricidus_tsai1999.rds")
source("R/dario_model_fitting.R")
aphis_citricida_fitted <- fit_devmodels(temp = a.citricidus_tsai1999$temperature,
                                        dev_rate = a.citricidus_tsai1999$rate_development,
                                        model_name = "all") #might be a bit slow
## examine them visually to better choose on ecological criteria and not only on statistical fitting
plot_devmodels(temp = a.citricidus_tsai1999$temperature,
               dev_rate = a.citricidus_tsai1999$rate_development,
               fitted_parameters = aphis_citricida_fitted)

therm_bounds_moth <- thermal_suitability_bounds(fitted_parameters = cabbage_moth_fitted,
                                                model_name = "lactin2",
                                                suitability_threshold = 75)
}
