% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{available_models}
\alias{available_models}
\title{Available Models Table}
\format{
\subsection{\code{available_models}}{

A data.frame/tibble with 13 rows and 6 columns. This package contain the equation names of the models that are
able to be fitted with \code{\link[=fit_devmodels]{fit_devmodels()}} and the equivalent names in their respective
source packages underlying the start value exploration function and formulas used in-code to fit models.
These models come from two packages that are already published:
\if{html}{\out{<a href="https://github.com/frareb/devRate">}}\code{devRate} \if{html}{\out{</a>}} and \if{html}{\out{<a href="https://github.com/padpadpadpad/rTPC">}}\code{rTPC} \if{html}{\out{</a>}}. \if{html}{\out{</br>}}

\describe{
\item{model_name}{This is the name of the model equations that the user can use as input in the package functions.}
\item{package}{names of the packages used by \code{\link[=fit_devmodels]{fit_devmodels()}} to obtain appropriate start values for the user-providen data.
When the package is \if{html}{\out{<a href="https://github.com/padpadpadpad/rTPC">}}rTPC package\if{html}{\out{</a>}}, start values are
automatically computed with \code{\link[rTPC:get_start_vals]{rTPC::get_start_vals()}}, which in turn relies on \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}. When the package is  \if{html}{\out{<a href="https://github.com/frareb/devRate">}}devRate package\if{html}{\out{</a>}},
iterative starting values are computed using \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}, using the parameters published in
\code{\link[devRate:devRateEqStartVal]{devRate::devRateEqStartVal()}} as first attempts to iterate. As an exception, if \code{model_name == "briere1"},
generic starting values are provided and advised to the user due to unrealistic value of some parameters in the
\link{devRate} data set.}
\item{source_model_name}{function name in source packages used in the background of this package}
\item{formula, working formula, params_formula}{formulas allowing automation of model fitting, with no information of interest for the user}
}
}
}
\source{
{\if{html}{\out{<b>}}Rebaudo, F., Struelens, Q., and Dangles, O. (2018)\if{html}{\out{</b>}}. Modelling temperature-dependent development rate and phenology in
arthropods: The \code{devRate} package for r. Methods Ecol Evol. 9: 1144-1150. \url{https://doi.org/10.1111/2041-210X.12935}

\if{html}{\out{<b>}}Padfield, D., O´Sullivan, H., and Pawar, S., (2021)\if{html}{\out{</b>}}. \code{rTPC} and \code{nls.multstart}: a new pipeline to fit thermal
performance curves in r. Methods Ecol Evol. 12: 1138-1143 \url{https://doi.org/10.1111/2041-210X.13585}}.\if{html}{\out{</br>}}
}
\usage{
data("available_models")
}
\description{
This data.frame, called \code{available_models} in your environment once the package is loaded,
contains the available models to fit using \code{\link[=fit_devmodels]{fit_devmodels()}}. It is useful for the user to know it
before applying functions in the package.
}
\keyword{datasets}
