% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_devmodels.R
\name{fit_devmodels}
\alias{fit_devmodels}
\title{Fit nonlinear regression models to development rate data}
\usage{
fit_devmodels(temp = NULL, dev_rate = NULL, model_name = NULL)
}
\arguments{
\item{temp}{a vector containing temperature treatments (predictor variable), must have at least three different temperature treatments}

\item{dev_rate}{a vector containing development rate estimates (1/days of development); must be of same length than temp}

\item{model_name}{"all" or alternatively one or several of the models listed in available_models}
}
\value{
this function returns a tibble with estimate and standard error for each parameter of the selected models and the AIC
}
\description{
Fit nonlinear regression models to development rate data
}
\examples{
data(p.xylostella_liu2002)

cabbage_moth_fitted <- fit_devmodels(temp = p.xylostella_liu2002$temperature,
                                        dev_rate = p.xylostella_liu2002$rate_development,
                                        model_name = c("all")) #might be a bit slow

}
