% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_risk.R
\name{map_risk}
\alias{map_risk}
\title{Map pest risk}
\usage{
map_risk(
  t_vals = NULL,
  t_rast = NULL,
  region = NULL,
  res = 2.5,
  path = getwd(),
  mask = TRUE
)
}
\arguments{
\item{t_vals}{Numeric vector of length 2 specifying the left (minimum) and right (maximum) temperature bounds for optimal performance of the target species.}

\item{t_rast}{Optional 12-layer SpatRaster with monthly mean temperatures for (at least) the target 'region'. If not provided, global WorldClim raster layers will be (down)loaded and cropped to 'region'.}

\item{region}{Optional object specifying the region to map. Can be a 'SpatVector' polygon map (obtained with 'terra::vect()'); or a 'SpatExtent' object (obtained with 'terra::ext()'), or a numeric vector of length 4 specifying the region coordinates in the order xmin, xmax, ymin, ymax; or a character vector of country name(s) in English. If NULL, the output maps will cover the entire 't_rast' if provided, or the entire world otherwise.}

\item{res}{Argument to pass to geodata::worldclim_global() specifying the spatial resolution for the raster maps to download, if 't_rast' is not provided. The default is 2.5 arc-minutes.}

\item{path}{Argument to pass to geodata::worldclim_global() (if 't_rast' is not provided) and/or to geodata::world() (if 'region' is a vector of country names) specifying the folder path for the downloaded maps. The default is the working directory.}

\item{mask}{Logical value to pass to terra::crop() specifying whether the output raster maps should be masked with the borders of the target 'region', if this is a polygon map or a vector of country names. The default is TRUE. If FALSE, the entire rectangular extent of 'region' will be used.}
}
\value{
This function returns a SpatRaster object with 13 raster layers of temperature (one for each month, and a final one with the mean value across months), with values only within the bounds of the input 't_vals'.
}
\description{
Map pest risk
}
\examples{
 risk_rasters <- map_risk(t_vals = c(12.5, 21.4), region = "Spain")
terra::plot(risk_rasters)
terra::plot(risk_rasters[[13]])
# terra::writeRaster(risk_rasters[[13]], "risk_mean.tif")

}
