% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateRpackage-package.R
\docType{package}
\name{mappestRisk-package}
\alias{mappestRisk}
\alias{mappestRisk-package}
\title{mappestRisk: Create Maps Forecasting Risk of Pest Occurrence}
\description{
The package provides maps that forecast crop pest risk of occurrence in response to experienced temperature. It takes as input development data parameterised in controlled conditions (i.e. constant or fluctuating temperatures) for different life stages of a species. The output can either assess climatic suitability or phenological progress simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/mappestRisk}
  \item Report bugs at \url{https://github.com/EcologyR/mappestRisk/issues}
}

}
<<<<<<< HEAD
\author{
\strong{Maintainer}: Darío dario.ssm2@gmail.com San-Segundo Molina (\href{https://orcid.org/0000-0002-7831-9623}{ORCID})

Authors:
\itemize{
  \item A. Marcia Barbosa
}

Other contributors:
\itemize{
  \item Antonio Jesús ajpelu@gmail.com Pérez-Luque (\href{https://orcid.org/0000-0002-1747-0469}{ORCID}) [conservator]
}

}
=======
>>>>>>> b1f05c53734c45665a149a9fe48ea0adc9306318
\keyword{internal}
