% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devmodels.R
\name{plot_devmodels}
\alias{plot_devmodels}
\title{Plot predictions of fitted thermal performance curves to your data}
\usage{
plot_devmodels(
  temp,
  dev_rate,
  fitted_parameters,
  species = NULL,
  life_stage = NULL
)
}
\arguments{
\item{temp}{a vector containing temperature treatments (predictor variable).
It must have at least four different temperature treatments. It must be numeric
and not containing NAs.}

\item{dev_rate}{a vector containing development rate estimates, calculated as
the reciprocal of days of development at each temperature (i.e., 1/days of development).
It must be numeric and of same length as \code{temp}.}

\item{fitted_parameters}{a \code{tibble} obtained with \code{fit_devmodels()} function including parameter names,
estimates, se, AICs and \if{html}{\out{<nls>}} objects (i.e. fitted_models) from \code{fit_devmodels()}
under \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}} approach.}
}
\value{
takes the fitted parameters table from \code{fit_devmodels()} and returns a plot
with the predicted values -i.e. development rate- across temperatures each for the
models that adequately converged. Facets of the resulting plots are automatically
sorted by lowest AIC values in descending order and additional information such
as number of parameters is displayed. Being a \if{html}{\out{<ggplot>}} object, it can be asigned to a user-
defined object.
}
\description{
The function\code{plot_devmodels()} displays a faceted plot of fitted TPCs
to help the user to select an appropriate model based both on statistical and ecological criteria.
}
\examples{
data("b.schwartzi_satar2002")

fitted_tpcs_bschwartzi <- fit_devmodels(temp = b.schwartzi_satar2002$temperature,
                                      dev_rate = b.schwartzi_satar2002$rate_value,
                                      model_name = "all")

plot_devmodels(temp = b.schwartzi_satar2002$temperature,
               dev_rate = b.schwartzi_satar2002$rate_value,
               fitted_parameters = fitted_tpcs_bschwartzi,
               species = "Brachycaudus swartzi",
               life_stage = "Nymphs")
}
\seealso{
\code{fit_devmodels()} for fitting Thermal Performance Curves to development rate data
}
