% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uncertainties.R
\name{plot_uncertainties}
\alias{plot_uncertainties}
\title{Draw bootstrapped TPCs to visualize uncertainty bands in parameter estimation of TPC fitting}
\usage{
plot_uncertainties(
  bootstrap_uncertainties_tpcs,
  temp,
  dev_rate,
  species,
  life_stage
)
}
\arguments{
\item{bootstrap_uncertainties_tpcs}{a \code{tibble} object output by \code{\link[=predict_curves]{predict_curves()}}
with as many curves -TPCs- as the number of iterations provided in \code{n_boots_samples} argument
therein whenever \code{propagate_uncertainty} was \code{TRUE}. Each TPC in this \code{tibble}  consist of
a collection of predictions for a set of temperatures from \code{temp - 20} to \code{temp + 15} with a
resolution of 0.1ºC and a unique identifier called \code{iter}.
In addition to the uncertainty TPCs, the estimate TPC is also explicit in the output tibble.}

\item{temp}{a vector containing temperature treatments (predictor variable).
It must have at least four different temperature treatments. It must be numeric
and not containing NAs.}

\item{dev_rate}{a vector containing development rate estimates, calculated as
the reciprocal of days of development at each temperature (i.e., 1/days of development).
It must be numeric and of same length as \code{temp}.}

\item{species}{a string containing the name of the species in the study. The function
converts the string to a title in the \if{html}{\out{<ggplot>}} object in italics.}

\item{life_stage}{a string containing the life stage studied for this rate-temperature
relationship. The function converts the string to a subtitle in the \if{html}{\out{<ggplot>}} object.}
}
\value{
a \if{html}{\out{<ggplot>}} object containing the visual representation of the estimate TPC and the
bootstrapped uncertainty curves as a ribbon. Each model is represented in a facet, and data points
are also explicit.
}
\description{
Draw bootstrapped TPCs to visualize uncertainty bands in parameter estimation of TPC fitting
}
\examples{
data("b.schwartzi_satar2002")

fitted_tpcs_bschwartzi <- fit_devmodels(temp = b.schwartzi_satar2002$temperature,
                                      dev_rate = b.schwartzi_satar2002$rate_value,
                                      model_name = "all")

plot_devmodels(temp = b.schwartzi_satar2002$temperature,
               dev_rate = b.schwartzi_satar2002$rate_value,
               fitted_parameters = fitted_tpcs_bschwartzi,
               species = "Brachycaudus swartzi",
               life_stage = "Nymphs") #choose "briere2", "thomas" and "lactin2"

#3. Obtain prediction TPCs with bootstraps for propagating uncertainty:
tpc_preds_boots_bschwartzi <- predict_curves(temp = b.swartzi_satar2002$temperature,
                                             dev_rate = b.swartzi_satar2002$rate_value,
                                             fitted_parameters = fitted_tpcs_bswartzi,
                                             model_name_2boot = c("briere2", "thomas", "lactin2"),
                                             propagate_uncertainty = TRUE,
                                             n_boots_samples = 100)


head(tpc_preds_boots_bschwartzi)

#4. Plot bootstrapped curves:

plot_uncertainties(bootstrap_uncertainties_tpcs = tpc_preds_boots_bschwartzi,
                   temp = b.swartzi_satar2002$temperature,
                   dev_rate = b.swartzi_satar2002$rate_value,
                   species = "Brachycaudus schwartzi",
                   life stage = "Nymphs")

}
