% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therm_suit_bounds.R
\name{therm_suit_bounds}
\alias{therm_suit_bounds}
\title{Determine Thermal Boundaries for Optimal Performance Level}
\source{
The dataset used in the example was originally published in Satar & Yokomi (2022) under the CC-BY-NC license
}
\usage{
therm_suit_bounds(
  preds_tbl = NULL,
  model_name = NULL,
  suitability_threshold = 75
)
}
\arguments{
\item{preds_tbl}{a \code{tibble} object inherited from \code{\link[=predict_curves]{predict_curves()}}
containing as many thermal performance curves as the number of iterations provided in \code{n_boots_samples} argument
therein and whenever \code{propagate_uncertainty} was set to \code{TRUE}; or alternatively the
estimated TPC predictions if \code{propagate_uncertainty} was set to \code{FALSE}.
Each TPC consists of predictions for temperatures ranging from \code{temp - 20} to \code{temp + 15}
with a resolution of 0.1°C.}

\item{model_name}{a string with one or several TPC models of those fitted first in \code{fit_devmodels()} and predicted
next in \code{predict_curves()}. Setting \code{model_name = "all"} is not allowed in this function. If the user wants to
calculate thermal boundaries for several models predicted and/or bootstrapped in \code{predict_curves()}, typing manually
all desired model names within a vector will be required.}

\item{suitability_threshold}{A numeric value from 50 to 100 representing the quantile of the curve that provides the user-defined optimal performance.
For instance, setting \code{suitability_threshold} to 80 identifies the top 20\% (or quartile 80) of the maximum values of the development rate predicted by the chosen TPC model.
If \code{suitability_threshold} equals 100, the function returns the optimum temperature for development rate.}
}
\value{
A tibble with six columns:
\itemize{
\item \code{model_name}: A string indicating the selected TPC model used for projections.
\item \code{suitability}: A string indicating the suitability threshold in percentage (see \code{suitability_threshold}).
\item \code{tval_left}: A number representing the lower thermal boundary delimiting the suitable region of the TPC.
\item \code{tval_right}: A number representing the upper thermal boundary delimiting the suitable region of the TPC.
\item \code{pred_suit}: A number corresponding to the predicted development rate value determining the chosen quantile threshold of the maximum rate (i.e., suitability percentage of maximum rate).
\item \code{iter}: A factor determining the TPC identity from the bootstrapping procedure in \code{\link[=predict_curves]{predict_curves()}} function, or \code{estimate} when it represents the estimated TPC fitted in \code{\link[=fit_devmodels]{fit_devmodels()}}.
}
}
\description{
This function calculates thermal boundaries that define the suitable region of a Thermal Performance Curve (TPC) corresponding to a user-defined optimal performance level.
}
\examples{
data("b.schwartzi_satar2002")

fitted_tpcs_bschwartzi <- fit_devmodels(temp = b.schwartzi_satar2002$temperature,
                                        dev_rate = b.schwartzi_satar2002$rate_value,
                                        model_name = "all")

plot_devmodels(temp = b.schwartzi_satar2002$temperature,
               dev_rate = b.schwartzi_satar2002$rate_value,
               fitted_parameters = fitted_tpcs_bschwartzi,
               species = "Brachycaudus swartzi",
               life_stage = "Nymphs") #choose "briere2", "thomas" and "lactin2"

# Obtain prediction TPCs with bootstraps for propagating uncertainty:
tpc_preds_boots_bschwartzi <- predict_curves(temp = b.schwartzi_satar2002$temperature,
                                             dev_rate = b.schwartzi_satar2002$rate_value,
                                             fitted_parameters = fitted_tpcs_bschwartzi,
                                             model_name_2boot = c("briere2", "thomas", "lactin2"),
                                             propagate_uncertainty = TRUE,
                                             n_boots_samples = 100)

head(tpc_preds_boots_bschwartzi)

# Plot bootstrapped curves:

plot_uncertainties(bootstrap_uncertainties_tpcs = tpc_preds_boots_bschwartzi,
                   temp = b.schwartzi_satar2002$temperature,
                   dev_rate = b.schwartzi_satar2002$rate_value,
                   species = "Brachycaudus schwartzi",
                   life_stage = "Nymphs")

#5. Calculate Q80 thermal bounds

boundaries_bschwartzi <- therm_suitability_bounds(preds_tbl = tpc_preds_boots_bschwartzi,
                                                  model_name = "lactin2",
                                                  suitability_threshold = 80)
head(boundaries_bschwartzi)
}
\references{
Angilletta, M.J., (2006). Estimating and comparing thermal performance curves. \if{html}{\out{<i>}}J. Therm. Biol.\if{html}{\out{</i>}} 31: 541-545.
(for reading on model selection in TPC framework)

Padfield, D., O'Sullivan, H. and Pawar, S. (2021). \if{html}{\out{<i>}}rTPC\if{html}{\out{</i>}} and \if{html}{\out{<i>}}nls.multstart\if{html}{\out{</i>}}: A new pipeline to fit thermal performance curves in \code{R}. \if{html}{\out{<i>}}Methods Ecol Evol\if{html}{\out{</i>}}. 00: 1-6

Rebaudo, F., Struelens, Q. and Dangles, O. (2018). Modelling temperature-dependent development rate and phenology in arthropods: The \code{devRate} package for \code{R}. \if{html}{\out{<i>}}Methods Ecol Evol\if{html}{\out{</i>}}. 9: 1144-1150.

Satar, S. and Yokomi, R. (2002). Effect of temperature and host on development of \if{html}{\out{<i>}}Brachycaudus schwartzi\if{html}{\out{</i>}} (Homoptera: Aphididae). \if{html}{\out{<i>}}Ann. Entomol. Soc. Am.\if{html}{\out{</i>}} 95: 597-602.
}
\seealso{
\code{browseVignettes("rTPC")} for model names, start values searching workflows, and
bootstrapping procedures using both \code{\link[rTPC:get_start_vals]{rTPC::get_start_vals()}} and \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}

\code{\link[=fit_devmodels]{fit_devmodels()}} for fitting Thermal Performance Curves to development rate data, which is in turn based on \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}.
\code{\link[=predict_curves]{predict_curves()}} for bootstrapping procedure based on the above-mentioned \code{rTPC} vignettes.
}
