% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_suitability_bounds.R
\name{thermal_suitability_bounds}
\alias{thermal_suitability_bounds}
\title{Obtain thermal traits representing boundaries of optimal thermal suitability for pest risk occurrence assessment}
\usage{
thermal_suitability_bounds(
  fitted_parameters,
  model_name,
  suitability_threshold = NULL
)
}
\arguments{
\item{fitted_parameters}{a \code{tibble} obtained with \code{fit_devmodels()} function including parameter names,
estimates, se, AICs and gnls objects (i.e. fitted_models) from \code{fit_devmodels()}.}

\item{model_name}{one or several of the models in fitted_parameters.
Must belong to available models in \code{dev_model_table}.}

\item{suitability_threshold}{Quantile of the curve; e.g. 75 would give
the temperatures within the values at which development rate is in the
top 25\% of its possible values. Def. 50. Note that \code{suitability_threshold < 50} is not allowed.}
}
\value{
a tibble/data.frame with name of the model and two thermal traits representing the above-mentioned suitability boundaries
}
\description{
Similar to calculating thermal breadth, \code{thermal_suitability_bounds()} computes the temperature bounds delimiting the optimum performance
as defined by user's suitability threshold.
}
\examples{
data("h.vitripennis_pilkington2014")
homalodisca_fitted <- fit_devmodels(temp = h.vitripennis_pilkington2014$temperature,
                                    dev_rate = h.vitripennis_pilkington2014$rate_development,
                                    model_name = c("all"),
                                    variance_model = "exp") #might be a bit slow

## examine plots to decide an appropriate model
plot_devmodels(temp = h.vitripennis_pilkington2014$temperature,
               dev_rate = h.vitripennis_pilkington2014$rate_development,
               fitted_parameters = homalodisca_fitted)

thermal_suitability_bounds(fitted_parameters = homalodisca_fitted,
                           model_name = "lactin2", # <- seems realistic
                           suitability_threshold = 75)

}
